i forgot
honestly trunk is such a terrible all-in-one piece of shit that i'd rather just have an all-in-none
linter. i have been trying to fix my linters and my formatters for so long that i legitimately
forgot everything i've done in this commit. do not use trunk. just use dprint and eslint and call it
a day. you don't need all-in-one. you really don't. it's so terrible.